+++ Python/src/24 Git/05_create_new_branch.py
@@ -1,3 +1,4 @@
+import sys
 from single_step import s
 
 s('cd myrepo')
@@ -7,7 +8,10 @@ s('git checkout mybranch')
 s('echo "branch file no. 1" > branch1.txt')
 s('echo "branch file no. 2" > branch2.txt')
 s('echo "branch file no. 3" > branch3.txt')
-s('rm hello3.txt')
+if sys.platform == "win32":
+    s('del hello3.txt')
+else:
+    s('rm hello3.txt')
 s('git add .') 
 s('git commit -m "created a new branch"')
 s('git push --set-upstream origin mybranch')
